requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
bcrypt==4.1.2
# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-mock>=3.11.0
faker>=19.0.0
flake8==7.0.0
black==24.2.0
pre-commit==3.6.0